// management/frontend/src/api/schemas/roleCreate.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * Ego Collection Management
 * OpenAPI spec version: 0.1.0
 */

export interface RoleCreate {
  /**
     * @minLength 1
     * @maxLength 64
     */
  role_name: string;
  business_line_id?: string | null;
  sort?: number;
  is_active?: boolean;
  remark?: string | null;
}