// management/frontend/src/api/zod/workspace-roles/workspace-roles.ts
export const workspaceRolesCreateRoleBodyRoleNameMax = 64;

export const workspaceRolesCreateRoleBodySortDefault = 1;
export const workspaceRolesCreateRoleBodyIsActiveDefault = true;

export const WorkspaceRolesCreateRoleBody = zod.object({
  "role_name": zod.string().min(1).max(workspaceRolesCreateRoleBodyRoleNameMax),
  "business_line_id": zod.union([zod.uuid(),zod.null()]).optional(),
  "sort": zod.number().default(workspaceRolesCreateRoleBodySortDefault),
  "is_active": zod.boolean().default(workspaceRolesCreateRoleBodyIsActiveDefault),
  "remark": zod.union([zod.string(),zod.null()]).optional()
})
// ...
export const workspaceRolesUpdateRoleBodyRoleNameMax = 64;

export const workspaceRolesUpdateRoleBodySortDefault = 1;
export const workspaceRolesUpdateRoleBodyIsActiveDefault = true;

export const WorkspaceRolesUpdateRoleBody = zod.object({
  "role_name": zod.string().min(1).max(workspaceRolesUpdateRoleBodyRoleNameMax),
  "business_line_id": zod.union([zod.uuid(),zod.null()]).optional(),
  "sort": zod.number().default(workspaceRolesUpdateRoleBodySortDefault),
  "is_active": zod.boolean().default(workspaceRolesUpdateRoleBodyIsActiveDefault),
  "remark": zod.union([zod.string(),zod.null()]).optional()
})